Split walkers into separate files
Enables individual walkers to import required dependencies without
introducing circular dependencies for other walkers.
